Making a movie from vtk files in sample/VTK/<any name> subdirectory. Renaming subdirectory and files to sample/VTK/sample_NNN.vtk Changed images size. Draw axis with cm scale. Kmake_movie.sh makes movie on server from simulation results on K.

+++ visual_server/make_movie.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 echo "Removing dir sample/"
-rm -rf sample
+if [[ -d "sample" ]]; then
+    rm -rf sample
+fi
 echo "Making movie out.mp4"
-rm out.mp4
+if [[ -f "out.mp4" ]]; then
+    rm out.mp4
+fi
 ffmpeg -i export/img_%04d.png -vcodec mpeg4 -q:v 2 out.mp4
